requirements.txt
# FinGuard IntelliAgent - Python Dependencies
# Milestone 1 - Baseline Requirements

# ============================================================================
# Core Framework
# ============================================================================
fastapi==0.104.1              # Modern web framework for building APIs
uvicorn[standard]==0.24.0     # ASGI server for FastAPI
pydantic==2.5.0               # Data validation using Python type annotations
pydantic-settings==2.1.0      # Settings management with Pydantic

# ============================================================================
# Anthropic ADK & AI
# ============================================================================
anthropic==0.7.8              # Anthropic API client for Claude models
google-generativeai==0.3.1    # Google Gemini API client for RAG
# Note: anthropic-adk will be added in Milestone 2 when available

# ============================================================================
# Data Processing & Analysis
# ============================================================================
pandas==2.1.3                 # Data manipulation and analysis
numpy==1.26.2                 # Numerical computing
python-dateutil==2.8.2        # Date and time utilities

# ============================================================================
# SMS & Text Processing
# ============================================================================
regex==2023.10.3              # Advanced regex operations for SMS parsing
phonenumbers==8.13.26         # Phone number parsing and validation

# ============================================================================
# Database (prepared for Milestone 2)
# ============================================================================
# sqlalchemy==2.0.23          # SQL toolkit and ORM (uncomment in Milestone 2)
# psycopg2-binary==2.9.9      # PostgreSQL adapter (uncomment in Milestone 2)

# ============================================================================
# Environment & Configuration
# ============================================================================
python-dotenv==1.0.0          # Load environment variables from .env file
python-multipart==0.0.6       # Form data parsing for FastAPI

# ============================================================================
# HTTP & Networking
# ============================================================================
httpx==0.25.2                 # Async HTTP client
requests==2.31.0              # HTTP library for API calls

# ============================================================================
# Development & Testing
# ============================================================================
pytest==7.4.3                 # Testing framework
pytest-asyncio==0.21.1        # Async test support
pytest-cov==4.1.0             # Coverage plugin for pytest
black==23.11.0                # Code formatter
flake8==6.1.0                 # Linting tool
mypy==1.7.1                   # Static type checker
isort==5.12.0                 # Import sorting

# ============================================================================
# Utilities
# ============================================================================
python-json-logger==2.0.7     # JSON logging support
rich==13.7.0                  # Beautiful terminal formatting
typer==0.9.0                  # CLI application framework

# ============================================================================
# Jupyter & Notebooks (for testing & development)
# ============================================================================
jupyter==1.0.0                # Jupyter notebook
ipykernel==6.27.1             # IPython kernel for Jupyter
matplotlib==3.8.2             # Plotting library
seaborn==0.13.0               # Statistical visualization

# ============================================================================
# Security
# ============================================================================
cryptography==41.0.7          # Cryptographic recipes and primitives
python-jose[cryptography]==3.3.0  # JWT handling

# ============================================================================
# Notes for Future Milestones
# ============================================================================
# Milestone 2 additions will include:
# - anthropic-adk (when officially released)
# - Database drivers (SQLAlchemy, psycopg2)
# - Celery for background tasks
# - Redis for caching
# - Additional ML libraries as needed
